WD-306 Optimize utils with boilerplate
