+++ web-utils/build.gradle.kts
@@ -13,7 +13,7 @@ apply(plugin = "java")
 apply(plugin = "kotlin")
 
 group = "com.icerockdev"
-version = "0.4.0"
+version = "0.5.0"
 
 val sourcesJar by tasks.registering(Jar::class) {
     classifier = "sources"
